refcnt: make refcnt_t opaque
C compilers happily cast across typedefs.  For example, if we have:

	typedef foo_t bar_t;

	void test(bar_t *);

Even though the intention is to have test() take a bar_t pointer, the compiler
will happily let you pass in a foo_t pointer.  Wrapping the parent type in a
structure forces the compiler to error out.

To prevent accidental uses of atomic_*() functions on refcnt_t, we wrap the
atomic_t in a struct.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>