refcnt: generalize the static structure concept
The string code (struct str) couldn't use the generic getref/putref because
we wanted statically allocated and initialized strings.  These must never be
freed and so they required a flag check and other special casing in getref
and putref.  It turns out we can make that all generic.  Now, when we
instantiate the getref and putref functions, we also take a is-static
callback function that we consult before modifying the refcount.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>