str: preallocate "/" string
Since this is a relatively common string we can preallocate it and return it
instead of going to the heap.

One of the consumers of libjeffpc saw over 1% reduction in struct str
allocations during startup:

cache                        buf     buf     buf     buf  memory     alloc alloc
name                        size  in use  in ptc   total  in use   succeed  fail
------------------------- ------ ------- ------- ------- ------- --------- -----
str-cache (before)            24    5347       -    5577    132K    642691     0
str-cache (after)             24    5347       -    5577    132K    634687     0

Note: This consumer is a heavy user of strings.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>