sexpr: failed number lexing should fail the parsing
If we fail to lex a number, we should fail the parsing instead of just
terminating the lexing - which effectively truncates the input just before
the problem token.

We terminate the parsing by defining an ERROR token, returning it when we
fail to lex a number, and purposefully not handling it in the parser
grammar.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>