buffer: add a sinkhole buffer type
The idea is similar to /dev/null - that is, a buffer that can be written but
one that discards all the writes.  This kind of behavior is useful when an
API requires a buffer, but we are only interested in the number of bytes
written and not the contents themselves.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>