buffer: allow buffer truncation
This is a similar idea to the POSIX truncate() function.  It sets the used
buffer size to a certain value.  If the new size is larger than the old
size, the newly visible bytes are zeroed out.  If the new size is smaller
than the old size, the bytes at the end are "lost" but the underlying buffer
allocation remains unchanged.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>