scgisvc: parse the query string
Now that we have query string parsing code, we can parse the query string of
every SCGI request we get and not have the consumers bother.

If we fail to parse it, we terminate the request without even consulting the
consumer.  While this isn't ideal, it is good enough for now.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>