sexpr: reduce usage of NULL meaning empty cons cell
A NULL struct val pointer has been used (inconsistently) as an alias for an
empty cons cell.  This made sense before we supported preallocation of vals,
but today there is no reason to continue supporting this.  Unfortunately,
since this NULL pointer assumption permeates the library, we'll have to get
rid of it in phases.  This commit implements the first phase - reducing the
amount of NULLs generated instead of VT_CONS structs.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>