sexpr: rewrite s-expression dumping code to pass the tests
The test cases have the correct expected outputs, but the dumping code was
terribly broken.  This wasn't obvious until recently when the sexpr parser
test utility started checking the dump output against the expected output.
It was simpler to rewrite the dumping code from scratch instead of trying to
address the numerious issues with the previous version.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>