tree: stop inlining node finding in insertion code
If someone calls tree_insert_here() with a NULL cookie, they can expect an
O(lg n) search to find a place for the new node.  This is so much more
expensive than a single function call that it is better to just call the
full-fledged tree_find() function and reduce the size of tree_insert_here().

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>