scgisvc: send response body iff request method was not HEAD
A HEAD request made over a scgi socket should *not* return a body.  At least
this is what nginx expects given that the moment it receives the empty line
terminating the headers, it closes the connection to us.  Attempting to send
the body has been causing occasional -ESPIPE errors to be logged by blahgd.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>