synch: dynamically find circular lock class orderings & warn about them
As locks are used, keep track of observed lock class orderings in a directed
acyclic graph.  If a new ordering causes the graph to develop a cycle, the
new ordering is at odds with previously observed orderings and there is a
potential for a deadlock.

For each lock class, we keep track of its dependencies - lock classes found
at the top of the held locks stack when a lock of the original class was
being acquired.  Collectively, all the dependencies of all the used lock
classes make up the directed acyclic graph.

Note: Instead of keeping track of lock orderings, this code keeps track of
lock class orderings.  This is a much simpler (and lighter weight) task
making the runtime overhead rather low.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>