tests: support for tests that are expected to panic
So far we've been only able to check for behavior that did not panic the
process.  This commit changes that by allowing the test to set an expected
panic string.

If we encountered a panic which matches the expected panic string, the test
scaffolding will gracefully exit - making the test pass.  Unexpected panics
or a lack of a panic when we expected one will make the test fail.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>