int: fail all str2uXX calls with -ERANGE on negative looking input
We have to manually check for negative looking input since (1) negative
numbers cannot be represented with uintXX_t types, and (2) strtoull doesn't
fail (instead it does 2's complement negation of the unsigned bits).

We don't bother differentiating between input consisting of a negative but
otherwise valid number and a non-numeric string starting with a dash.
Ideally, we'd return -ERANGE for the former and -EINVAL for the latter, but
it isn't worth the additional clock cycles to differentiate between the two.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>