cmake: make libjeffpc's dependancies private
By default, all libraries listed in target_link_libraries are used when
linking the target *and* any target that depends on it.  In this case, it
meant that each of the test bins got linked against libexecinfo, etc.
Marking libjeffpc's deps as private will make cmake use them only when
linking libjeffpc but none of its consumers.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>