tests: explicitly specify test input file extension for file-based test
Instead of specifying a list of globs to expand as input file names, specify
a list of input directories and an input file extension to use in each dir.
This slightly reduces the clutter by not having to repeat the glob for each
directory, but more importantly it makes the next few commits simpler.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>