tests: read test input & output files generically
Instead of having each file-based test mangle the input file name to produce
the output file name and then doing I/O on it, do all that in the shared
generic test code.

If the expected extension is the empty string, the generic code skips output
file processing.  It is up to the individual test to do its own file name
mangling and I/O.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>