buffer: add buffer_init_heap to support stack allocated heap buffers
The underlying buffer is still heap allocated.

Sometimes, it is wasteful to allocate the buffer structure itself on the
heap only to free it shortly thereafter.  In those cases, we can use the
stack to hold the struct buffer (much like we do for the sink/static/etc.
buffers) while using the heap for the underlying buffer.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>